refactor: convert courses creation from modal to standalone page
